Provides optional arguments workaround
This commit allows users to omit the newup output directory argument. We
cannot set the argument as optional on the command itself in case
template authors want to specify one of their arguments/options as
required. This allows us to inject the current directory as the value if
it has been omitted assuming the package template does not specify any
options/arguments.

Related to commit:
https://github.com/newup/core/commit/4731e29e8a19b98cbf17bcfb9ccd55a8f2be3186